significantly speed up searching (by XX%)
Go through list twice.  Normally, this would be really slow.

BUT - exploit the fact that (usually), the data that we are looking for
is at the top or bottom.  Remember when we've successfully found the
data that we are looking for,

Search from the top for only the data that is usually at the top.
Search from the bottom only for the data that is usually at the bottom.

Only run the regexs for the data we haven't found.

Short-circuit when we've found all the data we need.

Cuts down on the number of regexes called by a factor of 5
(PanTiltBotCabling.gx.gcode used to use 21 M regexes;it now uses 4.7M)